remove redundent launch option
