+++ WheelWizard/Services/Launcher/Helpers/RetroRewindLaunchHelper.cs
@@ -17,7 +17,6 @@ public static class RetroRewindLaunchHelper
     {
         var language = (int)SettingsManager.RR_LANGUAGE.Get();
         var removeBlur = (bool)SettingsManager.REMOVE_BLUR.Get();
-        var gameMode = (int)SettingsManager.RETRO_REWIND_GAMEMODE.Get();
 
         var launchConfig = new LaunchConfig
         {
@@ -31,7 +30,7 @@ public static class RetroRewindLaunchHelper
                     {
                         Options = new[]
                         {
-                            new OptionConfig { Choice = gameMode, OptionName = "Pack", SectionName = "Retro Rewind" },
+                            new OptionConfig { Choice = 1, OptionName = "Pack", SectionName = "Retro Rewind" },
                             new OptionConfig { Choice = 2, OptionName = "My Stuff", SectionName = "Retro Rewind" },
                             new OptionConfig { Choice = language, OptionName = "Language", SectionName = "Retro Rewind" },
                             new OptionConfig { Choice = removeBlur ? 1 : 0, OptionName = "Remove Blur", SectionName = "Retro Rewind" }
