+++ WheelWizard/Services/Settings/SettingsManager.cs
@@ -84,10 +84,6 @@ public class SettingsManager
                                                             }).SetDependencies(DOLPHIN_COMPILATION_MODE, DOLPHIN_COMPILE_SHADERS_AT_START, DOLPHIN_MSAA, DOLPHIN_SSAA);
 
     private static RrGameMode _internalRrGameMode = RrGameMode.RETRO_TRACKS;
-    public static Setting RETRO_REWIND_GAMEMODE = new VirtualSetting(typeof(RrGameMode),
-                                                            value => _internalRrGameMode = (RrGameMode)value!,
-                                                            () => _internalRrGameMode
-                                                            );
     #endregion
