+++ WheelWizard/Views/Pages/HomePage.axaml.cs
@@ -22,8 +22,7 @@ public partial class HomePage : UserControl
     private static int _launcherIndex = 0; // Make sure this index never goes over the list index
     private static List<ILauncher> _launcherTypes = new List<ILauncher>()
     { // Add Launcher instances here , every launch instance is a new start screen
-        new RrLauncher(),
-        new RrCtsLauncher()
+        new RrLauncher()
         //GoogleLauncher.Instance
     };
